+++ scripts/moduleConfigs/fontsModules/nuxtFonts/nuxtConfig.ts
@@ -1,0 +1,9 @@
+export default {
+    fonts: {
+        provider: 'google',
+        defaults: {
+            weights: [400, 500, 600, 700, 800, 900],
+            styles: ['normal', 'italic'],
+        },
+    }
+}